Pálya backend teszteinek előkészítése
- helpers.js kiszervezése a test mappába, hogy lehessen használni a pálya teszteknél is
- Kommentekhez tesztek írásának elkezdése
