+++ backend/package.json
@@ -15,7 +15,8 @@
         "#mapcreatortest/*": "./test/mapcreator/*",
         "#root/*": "./*",
         "#mapcreator/*": "./api/mapcreator/*",
-        "#gamemaps/*": "./api/gamemaps/*"
+        "#gamemaps/*": "./api/gamemaps/*",
+        "#testhelpers/*": "./test/helpers/*"
     },
     "author": "Kardos Krisztián",
     "license": "ISC",
