+++ backend/test/gamemaps/comments.test.js
@@ -1,0 +1,44 @@
+const { createTestApp } = require("./helpers/setup-test.js");
+const {
+    randomId,
+    buildRequest,
+    testInvalidIDs,
+    testRequiresAuth
+} = require("#testhelpers/helpers.js");
+
+const ERRORS = require("#utils/error-messages.js");
+
+const database = require("#sql/database.js");
+const { mockConnection } = database;
+
+const requestWithSupertest = createTestApp();
+
+describe("Game Maps API - /api/game-maps/", () => {
+    describe("Comment Endpoints", () => {
+        describe("GET /:gameMapID/comments", () => {
+            const defaults = {
+                id: randomId()
+            };
+
+            const makeGetRequest = (overrides = {}) => buildRequest(
+                (id) => requestWithSupertest.get(`/api/game-maps/${encodeURIComponent(id)}/comments`),
+                overrides,
+                defaults
+            );
+
+            describe("Authorization (401, 403)", () => {
+                testRequiresAuth(() => makeGetRequest());
+            });
+
+            describe("Input validation (400, 413, 415, 422)", () => {
+                it("Should respond with 400 if the game map id is incorrect", async () => {
+                    await testInvalidIDs(
+                        (id) => makeGetRequest({ id }),
+                        ERRORS.GAMEMAP.INVALID_ID
+                    );
+                });
+            });
+
+        });
+    });
+});
